fix(iced): system info表示をモジュール分離
