+++ examples/iced-learn/src/main.rs
@@ -1,13 +1,14 @@
 use iced::widget::{
-    button, center, column, container, horizontal_space, row, slider, text, text_input, Column,
+    button, column, container, horizontal_space, row, slider, text, text_input, Column,
 };
-use iced::{border, system, Alignment, Element, Font, Task, Theme};
+use iced::{border, Alignment, Element, Font, Task, Theme};
+use scene::systeminfo::{self, SystemInfoMsg};
 
-const ICON_FONT: Font = Font::with_name("Consolas");
+pub mod scene;
 
 pub fn main() -> iced::Result {
     let s = iced::settings::Settings {
-        default_font: ICON_FONT,
+        default_font: Font::with_name("Consolas"),
         ..iced::settings::Settings::default()
     };
     iced::application(TitleView, State::update, State::view)
@@ -20,7 +21,7 @@ struct State {
     value: i64,
     slider: i32,
     content: String,
-    system_info: Option<system::Information>,
+    si: systeminfo::Scene,
 }
 
 impl Default for State {
@@ -30,7 +31,7 @@ impl Default for State {
             slider: 0,
             content: "test-app".to_string(),
             screen: Screen::Welcome,
-            system_info: None,
+            si: systeminfo::Scene::new(),
         }
     }
 }
@@ -38,7 +39,6 @@ impl Default for State {
 /// 表示内容の切り替え
 enum Screen {
     Welcome,
-    SystemInfo,
 }
 
 /// アプリケーションのタイトルを表示するための構造体
@@ -50,22 +50,8 @@ impl iced::application::Title<State> for TitleView {
         state.content.clone()
     }
 }
-
-#[derive(Debug, Clone)]
-enum SystemInfoMsg {
-    LoadSystemInfo,
-    SystemInfoLoaded(Box<system::Information>),
-}
-
-impl SystemInfoMsg {
-    fn loaded(info: system::Information) -> Message {
-        Message::SystemInfo(SystemInfoMsg::SystemInfoLoaded(Box::new(info)))
-    }
-}
-
 #[derive(Debug, Clone)]
 enum Message {
-    Welcome,
     SystemInfo(SystemInfoMsg),
     Increment,
     Decrement,
@@ -76,18 +62,7 @@ enum Message {
 impl State {
     fn update(&mut self, message: Message) -> Task<Message> {
         match message {
-            Message::SystemInfo(s) => match s {
-                SystemInfoMsg::LoadSystemInfo => {
-                    return system::fetch_information().map(SystemInfoMsg::loaded)
-                }
-                SystemInfoMsg::SystemInfoLoaded(info) => {
-                    self.system_info = Some(*info);
-                    self.screen = Screen::SystemInfo;
-                }
-            },
-            Message::Welcome => {
-                self.screen = Screen::Welcome;
-            }
+            Message::SystemInfo(s) => return self.si.update(s).map(Message::SystemInfo),
             Message::Increment => {
                 self.value += 1;
             }
@@ -102,85 +77,6 @@ impl State {
             }
         };
         Task::none()
-    }
-
-    fn system_info(info: &system::Information) -> Element<Message> {
-        use bytesize::ByteSize;
-
-        let content: Element<_> = {
-            let system_name = text!(
-                "System name: {}",
-                info.system_name.as_ref().unwrap_or(&"unknown".to_string())
-            );
-
-            let system_kernel = text!(
-                "System kernel: {}",
-                info.system_kernel
-                    .as_ref()
-                    .unwrap_or(&"unknown".to_string())
-            );
-
-            let system_version = text!(
-                "System version: {}",
-                info.system_version
-                    .as_ref()
-                    .unwrap_or(&"unknown".to_string())
-            );
-
-            let system_short_version = text!(
-                "System short version: {}",
-                info.system_short_version
-                    .as_ref()
-                    .unwrap_or(&"unknown".to_string())
-            );
-
-            let cpu_brand = text!("Processor brand: {}", info.cpu_brand);
-
-            let cpu_cores = text!(
-                "Processor cores: {}",
-                info.cpu_cores
-                    .map_or("unknown".to_string(), |cores| cores.to_string())
-            );
-
-            let memory_readable = ByteSize::b(info.memory_total).to_string();
-
-            let memory_total = text!(
-                "Memory (total): {} bytes ({memory_readable})",
-                info.memory_total,
-            );
-
-            let memory_text = if let Some(memory_used) = info.memory_used {
-                let memory_readable = ByteSize::b(memory_used).to_string();
-
-                format!("{memory_used} bytes ({memory_readable})")
-            } else {
-                String::from("None")
-            };
-
-            let memory_used = text!("Memory (used): {memory_text}");
-
-            let graphics_adapter = text!("Graphics adapter: {}", info.graphics_adapter);
-
-            let graphics_backend = text!("Graphics backend: {}", info.graphics_backend);
-
-            column![
-                system_name,
-                system_kernel,
-                system_version,
-                system_short_version,
-                cpu_brand,
-                cpu_cores,
-                memory_total,
-                memory_used,
-                graphics_adapter,
-                graphics_backend,
-                button("Back").on_press(Message::Welcome),
-            ]
-            .spacing(1)
-            .into()
-        };
-
-        center(content).into()
     }
 
     fn test_view(&self) -> Column<Message> {
@@ -229,13 +125,13 @@ impl State {
                     // button("Test").on_press(Message::LoadSystemInfo),
                     // button("Settings").on_press(Message::Settings),
                     self.test_view(),
+                    self.si.view(),
                 ]
                 .spacing(20)
                 .align_x(Alignment::Center);
 
                 content.into()
             }
-            Screen::SystemInfo => Self::system_info(self.system_info.as_ref().unwrap()),
         }
     }
 }
